+++ tests_compare_actual_vs_previous/test-compare_results_actual_vs_previous.R
@@ -15,8 +15,14 @@ meteo <- slb1_meteo[which(dates >= opts$startdate
 
 res_actual <- run_LWFB90(options_b90 = opts, param_b90 = parms, climate = meteo, soil =soil)
 
-meteo <- readRDS(system.file("testdata/meteo.rds", package = "LWFBrook90R"))
-res_previous <- readRDS(system.file("testdata/simresults.rds", package = "LWFBrook90R"))
+#meteo <- readRDS(system.file("testdata/meteo.rds", package = "LWFBrook90R"))
+#res_previous <- readRDS(system.file("testdata/simresults.rds", package = "LWFBrook90R"))
+
+meteo <- readRDS("tests_compare_actual_vs_previous/meteo.rds")
+res_previous <- readRDS("tests_compare_actual_vs_previous/simresults.rds")
+
+res_actual$output[,`:=`(snowlq = NULL, cc= NULL)]
+res_actual$layer_output[,`:=`(relawati = NULL)]
 
 test_that("sim-outputs of actual are equal to previous", {
   expect_equal(res_actual$output, res_previous$output)